northd: Call build_mcast_groups() from build_lflows()
As 'mcast_groups' and 'igmp_groups' need to be modified by
build_lflows(), we move build_mcast_groups() to build_flows(). This
is in preparation for a later commit which will create an lflow
node. If these structures form part of the input of the new lflow
we should not modify them as they would not form part of the mutable
state of that node.

Signed-off-by: Mark Gray <mark.d.gray@redhat.com>
Signed-off-by: Han Zhou <hzhou@ovn.org>