ovn-nbctl: Fix lrp-set-gateway-chassis.
Currently this command assumes that if the gateway_chassis record with
expected name exists it is set to the logical port, so once the record
is found it not set to the lrp again. However, this assumption is not
always true.

An example is that when combined with a lrp-del and then lrp-add
commands before lrp-set-gateway-chassis in the same transaction, the
gateway_chassis record will be found but it is not set to the lrp. This
is causing the gateway-chassis setting flapping in ovn-kubernetes'
cluster router logical ports.

This patch makes sure the gateway_chassis record (existed or newly
created) is set to the lrp's gateway-chassis column.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Acked-by: Numan Siddique <numans@ovn.org>
Acked-by: Mark Michelson <mmichels@redhat.com>