ovn-ic: physical: Support multicast_group flooding on IC transit switches.
Forwarding on logical patch ports that are part of a multicast group
normally happens on the local chassis.  With interconnect transit
switch, however, this doesn't work properly because patch ports in
the local AZ might not correspond to patch ports in the remote AZ.

This allows forwarding IP multicast traffic across different AZs when
the gateway routers in each AZ are configured to relay multicast and to
unconditionally flood on the gateway ports connected to the transit
switch.  This patch also adds a unit test to cover this scenario.

This is not really a feature, just a case that was missed due to the
fact that OVN-IC [0] and static IP multicast flood configuration [1] were
developed more or less at the same time.

[0] a3e4436c232 ("ovn-ic: Interconnection port controller.")
[1] 79308138891a ("ovn-northd: Add static IP multicast flood configuration")

Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=2028981
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>