vtep: provide option to check ovn-controller-vtep and ovn-northd versions
Simlar to ovn-controller's behavior with checking internal
version of ovn-northd and ovn-controller when option
ovn-match-northd-version is defined, this commit adds same
functionality for the ovn-controller-vtep daemon.

This configuration option is located in the configured
hardware_vtep database in the Global table's
other_config:ovn-match-northd-version column/key.
True value enforces check, while false or not defined values
skip the check.

The appropriate testcase is added as well.

Signed-off-by: Vladislav Odintsov <odivlad@gmail.com>
Signed-off-by: Numan Siddique <numans@ovn.org>