ovn-northd: Don't log transaction failures on standby instances.
We still need to try to ovsdb_idl_loop_commit_and_wait() on instances
that are in standby mode.  That's because they need to try to take the
lock.  But if they're in standby they won't actually build a transaction
json and will return early because they don't own the SB lock.

That's reported as an error by ovsdb_idl_loop_commit_and_wait() but we
shouldn't act on it.

Explicitly force a full recompute when transitioning from "standby" or
"paused" to "active".

Fixes: 953832eb17f3 ("ovn-northd: Don't trigger recompute for transactions that are in progress.")
Fixes: 4597317f16d1 ("northd: Introduce incremental processing for northd")
Reported-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Acked-by: Numan Siddique <numans@ovn.org>
Signed-off-by: Han Zhou <hzhou@ovn.org>
(cherry picked from commit 71f87a61e3dbbf7850055d027c1e5bccffa9de61)