ci: ovn-kubernetes: Generate DB model based on current OVN tree.
Make sure ovn-kubernetes is compiled against the NB/SB schema from the
OVN source tree.  By default upstream ovn-kubernetes uses the schema
from the latest released OVN version.  But that doesn't work fine if
in the meantime there were changes to the schema that are not yet part
of a stable release.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>