ovn-ic: Support IGMP/MLD in multi-AZ deployments.
Commit 974618c61de8 ("ovn-ic: physical: Support multicast_group flooding
on IC transit switches.") added support for statically forwarding IP
multicast traffic between AZs.  While this works, it's a waste of
resources to flood all IP multicast traffic to all instances of the
transit switch in all AZs.

Instead, we now extend the OVN IGMP/MLD support, in order to make it
function across availability zones.

Until now, all IGMP/MLD control packets were punted to ovn-controller
in the ingress pipeline of the logical switch that had multicast
snooping enabled.  We change this behavior for transit switches and
punt the packets both in the ingress pipeline (in the source AZ) but
also in the egress pipeline for remote ports (in the destination AZ).

This ensures that both OVN deployments (in the source and destination
AZs) dynamically learn the same IGMP groups on the transit switch.
There is, however, a catch: there's no guarantee that multicast groups
are learnt in the same order on all AZs so they might end up having
different tunnel keys.  In order to avoid the potential mismatch between
local and remote AZ multicast group tunnel keys we change multicast
group processing for "remote" ports and perform the group expansion on
the source node.

In order for the end-to-end solution to work, and for IP multicast
traffic to flow between logical switch ports that are behind gateway
routers in different AZs we also fix the logical router port static
multicast flooding (options:mcast_flood) implementation as it was always
dropping IGMP/MLD packets (these are always generated with TTL=1).

Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=2042952
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>