northd: Properly warn for NAT on LR with multiple gw ports.
This is not supported and ovn-northd was not logging it properly.
The previous attempt tried to log in init_nat_entries() but there
the logical router port configuration has not yet been parsed.

Move the check and log instead to where we install the NAT flows.
Also add a test.

Fixes: 15348b7b806f ("ovn-northd: Multiple distributed gateway port support.")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>