ovn-controller: Handle SB_Global:options:northd_internal_version in I-P engine.
Add an engine node en_northd_internal_version as an input to
en_lflow_output. When this node is updated, it triggers a recompute for
en_lflow_output. This node adds SB_Global as its only input, and it is
updated only when SB_Global's options:northd_internal_version is updated.

In a later patch the northd_internal_version will be used in
en_lflow_output and impact flow generation.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Acked-by: Numan Siddique <numans@ovn.org>