tests: Make "check CoPP config" more reliable.
At one point in the test, we expect that the output of `ovn-nbctl
copp-list` will produce:

arp: meter0
igmp: meter0

However, the output is created by walking a hashmap, and the ordering of
the items in the hashmap appears not to always be the same. On many
builds, I see the test fail because the order of the meters is reversed
in the output (i.e. igmp appears before arp). The data is correct, but
the test is being too strict about the expected ordering of the results.

This commit fixes the issue by running the results through `sort` to
ensure the ordering will be consistent.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Numan Siddique <numans@ovn.org>