ofctrl.c: Include group changes to bundle.
Move the group changes to the same bundle as the OF flow changes.
The steps in ofctrl_put were:
- add groups
- add meters
- bundle
        - change flows
- delete groups
- delete meters

Now it becomes:
- add meters
- bundle
        - add groups
        - change flows
        - delete groups
- delete meters

It is required for a future change that needs all operations in a
bundle. Ideally we should add meters to the bundle as well but it is not
supported yet by OVS.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Acked-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Numan Siddique <numans@ovn.org>
(cherry picked from commit e501112131f446e131b24aac237d6e60d2dfa38a)
Signed-off-by: Mark Michelson <mmichels@redhat.com>