northd: Honor ct-snat-zone option for common case.
Commit 4deac4509 introduced the new "czone" variants for ct_dnat and
ct_snat. This makes it so that a single conntrack zone is used for both
DNAT and SNAT on the datapath. The common CT zone chosen in this commit
was the DNAT zone.

The ct-snat-zone option for logical routers makes it so that an SNAT
zone can be explicitly configured instead of having a zone selected at
random by OVN. The problem is that since the DNAT zone is used as the
common zone ID, it means that when SNAT is performed, the SNAT is not
performed in the configured ct-snat-zone.

The fix for this is to change to using the SNAT zone as the common zone
if ct-snat-zone is configured on the logical router. This way, SNAT will
use the configured zone as expected. DNAT will also use this zone when
possible.

Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=2061619

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Numan Siddique <numans@ovn.org>